Cleaned up the codebase. Organized listeners by adding more specific packages.
