+++ src/main/java/me/repayed/tagminigame/listeners/interaction/PlayerInteractListener.java
@@ -1,22 +1,22 @@
-package me.repayed.tagminigame.listeners;
+package me.repayed.tagminigame.listeners.interaction;
 
 import me.repayed.tagminigame.TagMinigame;
 import me.repayed.tagminigame.game.GameState;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
-import org.bukkit.event.block.BlockPlaceEvent;
+import org.bukkit.event.player.PlayerInteractEvent;
 
-public class PlayerPlaceBlockListener implements Listener {
+public class PlayerInteractListener implements Listener {
 
     private final TagMinigame tagMinigame;
 
-    public PlayerPlaceBlockListener(final TagMinigame tagMinigame) {
+    public PlayerInteractListener(final TagMinigame tagMinigame) {
         this.tagMinigame = tagMinigame;
     }
 
     @EventHandler
-    public void playerPlaceBlock(BlockPlaceEvent event) {
-        if (!event.getPlayer().isOp()) {
+    public void onPlayerInteract(PlayerInteractEvent event) {
+        if(!event.getPlayer().isOp()) {
             event.setCancelled(true);
         } else {
             if(this.tagMinigame.getGameArena().getGameState() == GameState.INGAME) {
@@ -24,4 +24,5 @@ public class PlayerPlaceBlockListener implements Listener {
             }
         }
     }
+
 }
